[build] Update checkstyle to v 8.29 (#123)
### Motivation

The dependabots update of the checkstyle version ( #110 ) failed due to changes in the way the config file needs to be structured.  This PR updates the checkstyle version, AND fixes the config file.

### Modifications

This commit updates the checkstyle version, the checkstyle maven plugin
version.

This also updates the checkstyle config file to account for changes to
checkstyle configuration since the old version.

This also updates the files that were in "violation" of the checkstyle rules.